refactor: Create structure in desktop/electron-app
