+++ ui/desktop/electron-app/src/helpers/app-updater.js
