+++ ui/desktop/electron-app/src/ipc/handlers.js
@@ -1,9 +1,9 @@
 const { shell } = require('electron');
 const isDev = require('electron-is-dev');
-const handle = require('./ipc-handler');
-const boundaryCli = require('./boundary-cli');
-const origin = require('./origin');
-const sessionManager = require('./session/manager');
+const handle = require('./ipc-handler.js');
+const boundaryCli = require('../cli/index.js');
+const origin = require('../origin/index.js');
+const sessionManager = require('../session/manager.js');
 
 /**
  * Returns the current runtime origin, which is used by the main thread to
@@ -23,9 +23,7 @@ handle('setOrigin', async (requestOrigin) => {
 /**
  * Resets the origin.
  */
-handle('resetOrigin', async (requestOrigin) => {
-  origin.resetOrigin();
-});
+handle('resetOrigin', async () => origin.resetOrigin());
 
 /**
  * Opens the specified URL in an external browser.  Only secure HTTPs URLs are
