+++ ui/desktop/electron-app/src/origin/index.js
